+++ CHANGELOG.markdown
@@ -1,5 +1,6 @@
 next
 ----
+* Added `RepresentableOf`: `RepresentableOf rep f` is equivalent to `(Representable f, Rep f ~ rep)`.
 * Added `imapRep`, `ifoldMapRep`, `itraverseRep` to make it easier to define representable `FunctorWithIndex`, `FoldableWithIndex`, `TraversableWithIndex` instances from the `lens` package.
 * Add `GHC.Generics`-based default implementation for `Data.Functor.Rep.Representable` instances
 * Add `Data.Functor.Rep.Representable` instances for `Backwards`, `Reverse`, and the datatypes in `GHC.Generics`.